Change Rust step with more consistent function signature
Now it is essentially equal to the C++ declaration of step.

+++ src/rust/v0_baseline/src/lib.rs
@@ -41,10 +41,10 @@ fn _step(r: &mut [f32], d: &[f32], n: usize) {
 // Do not mangle function name to make library linking easier
 #[no_mangle]
 // Raw pointers can be dereferenced only inside 'unsafe' sections, hence function is marked as unsafe
-pub unsafe extern "C" fn step(r_raw: *mut f32, d_raw: *const f32, n: usize) {
+pub unsafe extern "C" fn step(r_raw: *mut f32, d_raw: *const f32, n: i32) {
     // Wrap raw pointers into 'not unsafe' Rust slices
-    let d = std::slice::from_raw_parts(d_raw, n * n);
-    let mut r = std::slice::from_raw_parts_mut(r_raw, n * n);
+    let d = std::slice::from_raw_parts(d_raw, (n * n) as usize);
+    let mut r = std::slice::from_raw_parts_mut(r_raw, (n * n) as usize);
     // Evaluate Rust implementation of the step-function
-    _step(&mut r, d, n);
+    _step(&mut r, d, n as usize);
 }
